chore: update Dockerfile to temporarily remove nodejs and npm
The Alpine 3.20 version of nodejs and npm are not compatible with
Phoenix 1.7.18.

Currently not able to update to Alpine 3.21 due to an erlang and gcc 14
bug where poll is not detected.

This is fixed in the maint branched of erlang, waiting for next erlang
release - https://github.com/erlang/otp/issues/9211.